Revert "Add Upsell CTA - v2 checkout "
