+++ support-frontend/assets/helpers/redux/checkout/product/state.ts
@@ -3,7 +3,6 @@ import type {
 	OtherAmounts,
 	SelectedAmounts,
 } from 'helpers/contributions';
-import { getGlobal } from 'helpers/globalsAndSwitches/globals';
 import { detect as detectCountryGroup } from 'helpers/internationalisation/countryGroup';
 import type { IsoCurrency } from 'helpers/internationalisation/currency';
 import type { BillingPeriod } from 'helpers/productPrice/billingPeriods';
@@ -47,7 +46,7 @@ export const initialProductState: ProductState = {
 	productOption: 'NoProductOptions',
 	fulfilmentOption: 'NoFulfilmentOptions',
 	billingPeriod: 'Monthly',
-	productPrices: getGlobal('window.guardian.productPrices') ?? {},
+	productPrices: window.guardian.productPrices,
 	selectedAmounts: {
 		ONE_OFF: 0,
 		MONTHLY: 0,
@@ -65,6 +64,6 @@ export const initialProductState: ProductState = {
 		},
 	},
 	currency,
-	orderIsAGift: getGlobal('window.guardian.orderIsAGift') ?? false,
+	orderIsAGift: window.guardian.orderIsAGift,
 	startDate: formatMachineDate(new Date()),
 };
